% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{clique_part}
\alias{clique_part}
\title{Generate a clique partition of the graph G defined by the adjacency matrix A.}
\usage{
clique_part(A)
}
\arguments{
\item{A}{An adjacency matrix.}
}
\value{
A clique partition in a list.
}
\description{
Generate a clique partition of the graph G defined by the adjacency matrix A.
}
\examples{
clique_part(A)
}

