% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{clique_select}
\alias{clique_select}
\title{Compute the computational complexity associated with a clique partition.}
\usage{
clique_select(m, Clist, V, NN)
}
\arguments{
\item{m}{Size of the parition.}

\item{Clist}{A list of cliques.}

\item{V}{Number of vertices in the graph.}

\item{NN}{Number of cliques in the graph.}
}
\value{
The computational complexity associated with a clique partition.
}
\description{
Compute the computational complexity associated with a clique partition.
}
\examples{
#Internal function only
}

