% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copCS.R
\docType{package}
\name{copCS}
\alias{copCS}
\alias{copCS-package}
\title{copCS: A package for using IPS to compute the concentration matrix Q.}
\description{
The copCS package relies on the igraph package to enumerate the cliques of
a graph. The cliques are then partitioned into sets and IPS is carried out
within each clique partition.
}

