% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{IPS}
\alias{IPS}
\title{Compute the concetration matrix via IPS.}
\usage{
IPS(p, A, eps = 1e-08)
}
\arguments{
\item{p}{A correlation between 0 and 1.}

\item{A}{An adjacency matrix.}

\item{eps}{(optional) Convergence tolerance.}
}
\value{
The concentration matrix \code{Q}.
}
\description{
Compute the concetration matrix via IPS.
}
\examples{
source("http://www.biostat.umn.edu/~hodges/RPLMBook/Datasets/10_Slovenian_stomach_cancer/Slovenia_stomach_cancer_data.txt")
A       <- -Q
diag(A) <- 0
Q       <- IPS(0.5, A)

}
\seealso{
\code{\link{clique_part}}
}

